Refactor: 로그인/회원가입 페이지 구조 개선
